refactor: Логика закрытия модальных окон вынесена из modal-overlay и modal-wondow
