+++ src/components/modals/modal-window.jsx
@@ -1,4 +1,5 @@
 import ReactDOM from 'react-dom';
+import React from 'react';
 import style from './modal-window.module.css';
 import { CloseIcon } from '@ya.praktikum/react-developer-burger-ui-components';
 import PropTypes from 'prop-types';
@@ -6,6 +7,20 @@ import { ingredientPropType } from '../../utils/prop-types';
 import ModalOverlay from './modal-overlay';
 
 function ModalWindow(props) {     
+
+    React.useEffect(() => {
+        document.addEventListener('keydown', handleKey);
+        
+        return () => {
+            document.removeEventListener('keydown', handleKey);
+        }        
+    }, [])
+
+    const handleKey = (e) => {
+        if(e.key === 'Escape') {
+            props.handleCloseModal();
+        }
+    }
 
     return ReactDOM.createPortal((        
         <>
